requirements.txt
diffusers
torch
accelerate
transformers
hf_xet
fastapi
pillow
uvicorn
fastapi-cors
psycopg2-binary 
python-dotenv
bcrypt
python-multipart
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118